+++ NEWS.md
@@ -1,5 +1,6 @@
 # Version 0.1.2
 
+* Acceptance tests.
 * JSON responses are now returned with `Content-Type: application/json` 
   instead of `Content-Type: text/json`.